Add simple pagination for the home page.
This also makes a single call to the database to get all packages and retrieves the README information lazily, for a bit more efficiency. But it still needs to query, sort and filter the complete list of packages, so a future change will be needed to push this to the database layer.